More memory fixes
Fixed the leaking GameSettings* buff and the leaking Players* that the
Client expected the ClientController to clean up.

In order for one part of this, I also needed to change gamewindow.cpp to
expect an unsigned int, rather than an int

I believe the only other possible memory errors at this point are in
gui/gamewindow.cpp. I believe it is losing a ton of closures

+++ src/gui/gamewindow.cpp
@@ -113,7 +113,7 @@ void GameWindow::updatePlayer(unsigned int id) {
     if(p != NULL)
     {
         newString = (*p)->getName( );
-        if( (int) id == *myID_subscriber->getData<int>( ) )
+        if( id == *myID_subscriber->getData<unsigned int>( ) )
             newString += " (me)";
 
         if(id == 0)
